move some function defs from Interpreter to AbstractInterpreter
