+++ include/souper/Infer/AbstractInterpreter.h
@@ -3,6 +3,9 @@
 
 #include "llvm/Support/KnownBits.h"
 
+#include "souper/Inst/Inst.h"
+#include "souper/Infer/Interpreter.h"
+
 namespace souper {
   namespace BinaryTransferFunctionsKB {
     llvm::KnownBits add(const llvm::KnownBits &lhs, const llvm::KnownBits &rhs);
@@ -25,6 +28,27 @@ namespace souper {
     llvm::KnownBits ule(const llvm::KnownBits &lhs, const llvm::KnownBits &rhs);
     llvm::KnownBits sle(const llvm::KnownBits &lhs, const llvm::KnownBits &rhs);
   }
+
+  std::string knownBitsString(llvm::KnownBits KB);
+
+  bool isConcrete(souper::Inst *I,
+		  bool ConsiderConsts = true,
+		  bool ConsiderHoles = true);
+
+  llvm::KnownBits findKnownBits(Inst* I,
+				ValueCache& C,
+				bool PartialEval = true);
+  llvm::KnownBits findKnownBitsUsingSolver(Inst *I,
+					   Solver *S,
+					   std::vector<InstMapping> &PCs);
+
+  llvm::ConstantRange findConstantRange(souper::Inst* I,
+					souper::ValueCache& C,
+					bool PartialEval = true);
+  llvm::ConstantRange findConstantRangeUsingSolver(souper::Inst* I,
+						   Solver *S,
+						   std::vector<InstMapping> &PCs);
+
 }
 
 #endif
